refactor file structure into modular files
Restructured backend code into modular components to improve maintainability and separation of concerns.
